+++ doc/source/doxygen-docs/changelog.md
@@ -1,7 +1,9 @@
 \page changelog Change Log
 
 # Version 2.14.6: UNRELEASED
-(None yet)
+- Changes in libraries:
+  - \ref mrpt_opengl_grp:
+    - mrpt::opengl::Texture now supports rendering 16-bit images.
 
 # Version 2.14.5: Released Nov 4th, 2024
 - Upgraded nanoflann to v1.6.2 (fixes a potential bug).